+++ AtelieConfeiteira/ScheduleCake/agendabolo.API/agendabolo.API/Controllers/FabricantesController.cs
@@ -50,7 +50,7 @@ namespace Agendabolo.Controllers
 
         [HttpGet]
         [Route("{id}")]
-        public IActionResult get(ulong id)
+        public IActionResult get(int id)
         {
             try
             {
@@ -130,7 +130,7 @@ namespace Agendabolo.Controllers
 
         [HttpDelete]
         [Route("{id}")]
-        public IActionResult delete(ulong id)
+        public IActionResult delete(int id)
         {
             try
             {
