+++ AtelieConfeiteira/ScheduleCake/agendabolo.API/agendabolo.domain/Core/Ingredientes/IIngredienteRepository.cs
@@ -5,7 +5,7 @@ using Agendabolo.GenericRepository;
 
 namespace Agendabolo.Core.Ingredientes
 {
-    public interface IIngredienteRepository : IGenericRepository<IngredienteDTA, ulong>
+    public interface IIngredienteRepository : IGenericRepository<IngredienteDTA, int>
     {
     }
 }
