+++ AtelieConfeiteira/ScheduleCake/agendabolo.API/agendabolo.domain/Core/Ingredientes/IngredienteRepository.cs
@@ -5,7 +5,7 @@ using System.Linq.Expressions;
 
 namespace Agendabolo.Core.Ingredientes
 {
-    public class IngredienteRepository : GenericRepository.GenericRepositoryDbContext<IngredienteDTA, ulong>, IIngredienteRepository
+    public class IngredienteRepository : GenericRepository.GenericRepositoryDbContext<IngredienteDTA, int>, IIngredienteRepository
     {
         public IngredienteRepository(ApplicationDbContext context) 
             : base(context)
