+++ AtelieConfeiteira/ScheduleCake/agendabolo.API/agendabolo.domain/Core/Receitas/IReceitaRepository.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace Agendabolo.Core.Receitas
 {
-    public interface IReceitaRepository: IGenericRepository<ReceitaDTA, ulong>
+    public interface IReceitaRepository: IGenericRepository<ReceitaDTA, int>
     {
         void RemoveItems(IEnumerable<ReceitaIngredienteDTA> ingredientesReceita);
     }
